instruction counting and resolving comments

+++ utils.h
@@ -14,6 +14,6 @@ uint32_t u32_abs_i32(int32_t x);
 void uart_putdec_u32(uint32_t v);
 void uart_putdec_i32(int32_t v);
 
-void uart_puthex_u64(int32_t v);
+void uart_puthex_u64(uint64_t v);
 
 extern const int16_t sin_table[];
